feat: CartService
- Cart Repository를 위한 서비스 로직 작성

+++ kakao-shopping/src/main/java/com/kakao/shopping/dto/cart/CartInsertRequest.java
@@ -1,7 +1,7 @@
 package com.kakao.shopping.dto.cart;
 
 public record CartInsertRequest(
-        int optionId,
-        int quantity
+        Long optionId,
+        Long quantity
 ) {
 }
