+++ kakao-shopping/src/main/java/com/kakao/shopping/dto/cart/CartUpdateRequest.java
@@ -1,7 +1,7 @@
 package com.kakao.shopping.dto.cart;
 
 public record CartUpdateRequest(
-        int cartId,
-        int quantity
+        Long cartId,
+        Long quantity
 ) {
 }
